+++ lib/spy/spyOn.ts
@@ -1,17 +1,27 @@
 import { Spy } from './spy';
 import { MockRegistry } from '../mocking';
 
-export function spyOn(object: any, methodName: string) {
-  const originalMethod = object[methodName];
-  const spy = new Spy(originalMethod, object);
+export function spyOn(object: any, methodName?: string) {
+  const isFunctionSpy = typeof object === 'function' && !methodName;
 
-  object[methodName] = spy.spyFn;
+  const originalMethod = isFunctionSpy ? object : object[methodName!];
+  const spy = new Spy(originalMethod, isFunctionSpy ? null : object);
 
-  MockRegistry.registerMock({
-    restoreAll: () => {
-      object[methodName] = originalMethod;
-    },
-  });
+  if (isFunctionSpy) {
+    object = spy.spyFn;
+    MockRegistry.registerMock({
+      restoreAll: () => {
+        object = originalMethod;
+      },
+    });
+  } else {
+    object[methodName!] = spy.spyFn;
+    MockRegistry.registerMock({
+      restoreAll: () => {
+        object[methodName!] = originalMethod;
+      },
+    });
+  }
 
   return spy;
 }